requirements.txt
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
xgboost==2.0.3
shap==0.44.0
prophet==1.1.5
folium==0.15.1
streamlit==1.31.0
streamlit-folium==0.18.0
tweepy==4.14.0
googletrans==4.0.0rc1
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.1
geopandas==0.14.3
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.1
httpx==0.26.0
pydantic==2.5.3
deepface==0.0.93
tf-keras
opencv-python-headless==4.9.0.80
cryptography==42.0.2
scipy==1.12.0
Pillow==10.2.0
fpdf2==2.7.8
